+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## head
 
+* Add syntax highlighting via pygments.rb
+
 ## v0.0.1
 
 Initial public release